Adds some tests for opcode decoding and fixing issues with it
get_bytes! macro was broken for most things, and not worth the confusion
about it, since the only time variadic arguments are useful here is when
we don't need the size argument, so we can instead split it into two
functions: get_nibble and get_nibbles, of which the first is just a
shortcut meaning get_nibbles(location, 1). Also, renamed to use nibble
instead of byte, since that is what we are using for units in the
arguments, and what we want to extract from the opcodes.