Adds conditional compilation for wasm
Adds the wasm feature to chip8_core, allowing us to generate bindings
for parts of the core without effecting non-wasm compilation.

Also adds a Display impl for Instruction and some utility functions, for
usage on the frontend.

Renames the exposed wrapper on the wasm end, so that it's just Chip8